chore(behavior_path_planner): move scene module utils (#2927)
* chore(lane_change): move common utils


* chore(side_shift): move common utils


* chore(behavior_path_planner): move debug utils


---------
